Revert "mm: fix calculation of dirtyable memory"
This reverts commit 1f150924340c6d07f98623281fc5394060912ec3.

Commit 1f150924340c6d07f98623281fc5394060912ec3(from Linux 3.4.24)
introduced a weird bug where pages are not properly reclaimed after
freeing on our Samsung msm8974 kernel source tree.

Samsung is aware of this and does not follows this route with their
updated klte kernel source.

 - Reference 1 : https://github.com/CyanogenMod/android_kernel_samsung_klte/blob/cm-11.0/mm/page-writeback.c#L192
 - Reference 2 : https://github.com/CyanogenMod/android_kernel_samsung_klte/blob/cm-11.0/mm/page-writeback.c#L227
 - Reference 3 : https://github.com/CyanogenMod/android_kernel_samsung_klte/blob/cm-11.0/mm/page-writeback.c#L314

So let's follow the same.

Credit goes to hsk from http://www.matcl.com for testing and finding
the culprit of this bug.

Signed-off-by: arter97 <qkrwngud825@gmail.com>
Change-Id: I717b7107122b095ab0c07147f62876d16d0aa407